--- .env.development
@@ -1,4 +1,0 @@
-DATABASE_URL=postgresql://doadmin:l5al4hwte8qmj6x8@db-postgresql-sfo2-nextgen-do-user-1067699-0.db.ondigitalocean.com:25060/treetracker?ssl=no-verify
-DATABASE_SCHEMA=reporting
-NODE_LOG_LEVEL=debug
-NODE_ENV=development
