+++ package.json
@@ -41,8 +41,8 @@
     "eslint-plugin-prettier": "^3.2.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "husky": "^7.0.2",
-    "mocha": "^8.1.3",
-    "mock-knex": "^0.4.9",
+    "mocha": "^10.2.0",
+    "mock-knex": "^0.4.11",
     "nodemon": "^2.0.4",
     "nyc": "^15.1.0",
     "prettier": "^2.1.2",
@@ -56,7 +56,7 @@
     "eslint": "eslint --report-unused-disable-directives .",
     "eslint:fix": "npm run eslint -- --fix",
     "server": "nodemon server/server.js",
-    "test": "npm run test-integration;npm run test-repository",
+    "test": "npm run test-integration && npm run test-repository",
     "coverage": "nyc npm run test",
     "test-repository": "NODE_ENV=test mocha -r dotenv/config --exit ./server/repositories/**/*.spec.js",
     "test-integration": "NODE_ENV=test mocha -r dotenv/config --env dotenv_config_path=.env.test --exit --timeout 20000 --recursive './__tests__'",
